backend/requirements.txt
flask
werkzeug
pdfplumber
pillow
torch
transformers
